requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aioice==0.10.1
aiortc==1.13.0
aiosignal==1.3.2
alembic==1.16.2
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
asyncio-mqtt==0.16.2
attrs==25.3.0
av==14.4.0
babel==2.17.0
backrefs==5.9
bcrypt==4.3.0
black==25.1.0
CacheControl==0.14.3
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
coverage==7.9.1
cryptography==45.0.4
cycler==0.12.1
distlib==0.3.9
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.13
fastapi-cli==0.0.7
ffmpeg-python==0.2.0
filelock==3.18.0
firebase_admin==7.1.0
flake8==7.3.0
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2025.5.1
future==1.0.0
ghp-import==2.1.0
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.3.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.3
griffe==1.7.3
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
hyperframe==6.1.0
identify==2.6.12
idna==3.10
ifaddr==0.2.0
iniconfig==2.1.0
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
loguru==0.7.3
Mako==1.3.10
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mccabe==0.7.0
mdurl==0.1.2
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-autorefs==1.4.2
mkdocs-get-deps==0.2.0
mkdocs-material==9.6.14
mkdocs-material-extensions==1.3.1
mkdocstrings==0.29.1
mkdocstrings-python==1.16.12
mpmath==1.3.0
msgpack==1.1.1
multidict==6.5.0
mypy==1.16.1
mypy_extensions==1.1.0
networkx==3.5
-e git+https://github.com/SpherCodes/nexguard-surveillance.git@06b74afdf7992af7a37773df3a88792eca864d9b#egg=nexguard
nodeenv==1.9.1
numpy==2.3.0
opencv-python==4.11.0.86
orjson==3.10.18
packaging==25.0
paginate==0.5.7
paho-mqtt==2.1.0
pandas==2.3.0
passlib==1.7.4
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.6.0
pre_commit==4.2.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
psycopg2==2.9.10
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==2.22
pydantic==2.11.7
pydantic-extra-types==2.10.5
pydantic-settings==2.9.1
pydantic_core==2.33.2
pyee==13.0.0
pyflakes==3.4.0
Pygments==2.19.1
PyJWT==2.10.1
pylibsrtp==0.12.0
pymdown-extensions==10.16
pyOpenSSL==25.1.0
pyparsing==3.2.3
pytest==8.4.1
pytest-asyncio==1.0.0
pytest-cov==6.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
pyyaml_env_tag==1.1
requests==2.32.4
rich==14.0.0
rich-toolkit==0.14.7
rsa==4.9.1
scipy==1.15.3
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
sympy==1.14.0
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
ujson==5.10.0
ultralytics==8.3.156
ultralytics-thop==2.0.14
urllib3==2.5.0
uvicorn==0.34.3
virtualenv==20.31.2
watchdog==6.0.0
watchfiles==1.1.0
websockets==15.0.1
win32_setctime==1.2.0
yarl==1.20.1
